Removing all citizen related code and related references from skills service
