--- DFC.App.SkillsHealthCheck.Services.SkillsCentral/Messages/UpdateSkillsDocumentRequest.cs
@@ -1,11 +1,0 @@
-﻿using DFC.App.SkillsHealthCheck.Services.SkillsCentral.Models;
-
-namespace DFC.App.SkillsHealthCheck.Services.SkillsCentral.Messages
-{
-    public class UpdateSkillsDocumentRequest
-    {
-        public SkillsDocument SkillsDocument { get; set; }
-
-        public long DocumentId { get; set; }
-    }
-}
