--- DFC.App.SkillsHealthCheck.Services.SkillsCentral/Messages/UpdateSkillsDocumentResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace DFC.App.SkillsHealthCheck.Services.SkillsCentral.Messages
-{
-    public class UpdateSkillsDocumentResponse : GenericResponse
-    {
-    }
-}
