--- DFC.App.SkillsHealthCheck.Services.SkillsCentral/Models/CitizenIdentity.cs
@@ -1,11 +1,0 @@
-﻿namespace DFC.App.SkillsHealthCheck.Services.SkillsCentral.Models
-{
-    public class CitizenIdentity
-    {
-        //public PersonalDetails PersonalDetails { get; set; }
-
-        public ContactDetails ContactDetails { get; set; }
-
-        public MarketingPreferences MarketingPreferences { get; set; }
-    }
-}
