+++ DFC.App.SkillsHealthCheck/Startup.cs
@@ -118,7 +118,6 @@ namespace DFC.App.SkillsHealthCheck
                 return svc;
             });
             services.AddTransient<ISkillsHealthCheckService, SkillsHealthCheckService>();
-            services.AddTransient<IUserAssetService, UserAssetService>();
             services.AddTransient<IYourAssessmentsService, YourAssessmentsService>();
             services.AddTransient<IQuestionService, QuestionService>();
             services.AddTransient<IGovUkNotifyClientProxy, GovUkNotifyClientProxy>();
